--- backend/src/resource/asset/index-ec4ef66164c33b9c.css
@@ -1,23 +1,0 @@
-.hero.has-background {
-  position: relative;
-  overflow: hidden;
-}
-.hero-background {
-  position: absolute;
-  object-fit: cover;
-  object-position: bottom;
-  width: 100%;
-  height: 100%;
-}
-.hero-background.is-transparent {
-  opacity: 0.5;
-}
-
-.burger {
-  background-color: transparent;
-  border: none;
-}
-
-.navbar-brand {
-  align-items: center;
-}
