--- backend/src/resource/asset/snippets/blog-frontend-bec5c12e8e7a4cdd/asset/show.js
@@ -1,9 +1,0 @@
-export function showNotificationBox() {
-    document.getElementById('notification').style.display = 'block';
-}
-export function hideNotificationBox(event) {
-    let source = event.target || event.srcElement;
-    while (source.id !== 'notification' && source.parentNode)
-        source = source.parentNode;
-    source.style.display = 'none';
-}
