+++ backend/src/resource/page/index.html
@@ -15,9 +15,9 @@
     
     
     
-    <link rel="stylesheet" href="/asset/index-ec4ef66164c33b9c.css">
+    <link rel="stylesheet" href="/asset/index-ed6ced7a01113d2b.css">
 
-<link rel="preload" href="/asset/index-d3a025b611e964d_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
-<link rel="modulepreload" href="/asset/index-d3a025b611e964d.js"></head>
+<link rel="preload" href="/asset/index-c3ff3c2c98bba7d3_bg.wasm" as="fetch" type="application/wasm" crossorigin="">
+<link rel="modulepreload" href="/asset/index-c3ff3c2c98bba7d3.js"></head>
 <body>
-<script type="module">import init from '/asset/index-d3a025b611e964d.js';init('/asset/index-d3a025b611e964d_bg.wasm');</script></body></html>
+<script type="module">import init from '/asset/index-c3ff3c2c98bba7d3.js';init('/asset/index-c3ff3c2c98bba7d3_bg.wasm');</script></body></html>
