+++ backend/src/service/asset_list.rs
@@ -4,16 +4,14 @@
 ("codemirror.min.css", include_bytes!(r#"..\resource\asset/codemirror.min.css.gz"#), "text/css"),
 ("editor.html", include_bytes!(r#"..\resource\asset/editor.html.gz"#), "text/html; charset=utf-8"),
 ("fontawesome.min-5e9e696c59c57e83.css", include_bytes!(r#"..\resource\asset/fontawesome.min-5e9e696c59c57e83.css.gz"#), "text/css"),
-("index-d3a025b611e964d.js", include_bytes!(r#"..\resource\asset/index-d3a025b611e964d.js.gz"#), "text/javascript"),
-("index-d3a025b611e964d_bg.wasm", include_bytes!(r#"..\resource\asset/index-d3a025b611e964d_bg.wasm.gz"#), "application/wasm"),
-("index-ec4ef66164c33b9c.css", include_bytes!(r#"..\resource\asset/index-ec4ef66164c33b9c.css.gz"#), "text/css"),
+("index-c3ff3c2c98bba7d3.js", include_bytes!(r#"..\resource\asset/index-c3ff3c2c98bba7d3.js.gz"#), "text/javascript"),
+("index-c3ff3c2c98bba7d3_bg.wasm", include_bytes!(r#"..\resource\asset/index-c3ff3c2c98bba7d3_bg.wasm.gz"#), "application/wasm"),
+("index-ed6ced7a01113d2b.css", include_bytes!(r#"..\resource\asset/index-ed6ced7a01113d2b.css.gz"#), "text/css"),
 ("index.css", include_bytes!(r#"..\resource\asset/index.css.gz"#), "text/css"),
 ("logo.png", include_bytes!(r#"..\resource\asset/logo.png.gz"#), ""),
 ("regular.min-a0c258fb7c5f655d.css", include_bytes!(r#"..\resource\asset/regular.min-a0c258fb7c5f655d.css.gz"#), "text/css"),
 ("snippets/blog-frontend-0b390d584a933dd3/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-0b390d584a933dd3\asset\editor.js.gz"#), "text/javascript"),
 ("snippets/blog-frontend-0b390d584a933dd3/asset/show.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-0b390d584a933dd3\asset\show.js.gz"#), "text/javascript"),
-("snippets/blog-frontend-bec5c12e8e7a4cdd/asset/editor.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-bec5c12e8e7a4cdd\asset\editor.js.gz"#), "text/javascript"),
-("snippets/blog-frontend-bec5c12e8e7a4cdd/asset/show.js", include_bytes!(r#"..\resource\asset/snippets\blog-frontend-bec5c12e8e7a4cdd\asset\show.js.gz"#), "text/javascript"),
 ("solid.min-70c2e5caa950974d.css", include_bytes!(r#"..\resource\asset/solid.min-70c2e5caa950974d.css.gz"#), "text/css"),
 ("toastui-editor-all.min.js", include_bytes!(r#"..\resource\asset/toastui-editor-all.min.js.gz"#), "text/javascript"),
 ("toastui-editor.min.css", include_bytes!(r#"..\resource\asset/toastui-editor.min.css.gz"#), "text/css"),
